refactor:
Refactored test classes
